// src/services/types/create-media-presigned-url-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

import type {MediaFileResponse} from "./media-file-response.ts";

export type CreateMediaPresignedUrlResponse = {
	/**
	 * @description 업로드 HTTP 메서드
	 * @type string
	 */
	method: string;
	/**
	 * @description 업로드 Presigned URL
	 * @type string
	 */
	presignedUrl: string;
	/**
	 * @description 만료 시각
	 * @type string, date-time
	 */
	expiresAt: string;
	/**
	 * @description 파일 메타 정보
	 * @type object
	 */
	file: MediaFileResponse;
};